middleware module (ex. server) added to node start app for prod
